refactor: split doDownload into focused modules
Extract stream routing, result processing, cleanup, completion side
effects, and job finalization out of downloadExecutor.js (~1500 lines)
into five modules under modules/download/. Logic moves without
intended behavior changes. jobModule calls downloadCleanup directly
instead of instantiating a DownloadExecutor. Executor tests moved to
the new modules' suites.

Refs: #645
