+++ server/modules/__tests__/jobModule.test.js
@@ -12,11 +12,9 @@ jest.mock('../../models/video');
 jest.mock('../../models/jobvideo');
 jest.mock('../../models/jobvideodownload');
 jest.mock('../../models/channelvideo');
-jest.mock('../download/downloadExecutor', () => {
-  return jest.fn().mockImplementation(() => ({
-    cleanupInProgressVideos: jest.fn().mockResolvedValue()
-  }));
-});
+jest.mock('../download/downloadCleanup', () => ({
+  cleanupInProgressVideos: jest.fn().mockResolvedValue()
+}));
 jest.mock('../../logger');
 
 const { v4: uuidv4 } = require('uuid');
@@ -324,14 +322,12 @@ describe('JobModule', () => {
       JobVideoDownload.findAll.mockResolvedValue([]);
       JobVideoDownload.destroy.mockResolvedValue(0);
 
-      // Reset and reconfigure the DownloadExecutor mock for this specific test
+      // Reset and reconfigure the downloadCleanup mock for this specific test
       const mockCleanup = jest.fn().mockResolvedValue();
       jest.resetModules();
-      jest.doMock('../download/downloadExecutor', () => {
-        return jest.fn().mockImplementation(() => ({
-          cleanupInProgressVideos: mockCleanup
-        }));
-      });
+      jest.doMock('../download/downloadCleanup', () => ({
+        cleanupInProgressVideos: mockCleanup
+      }));
 
       JobModule = require('../jobModule');
